+++ fastly/data_source_compute_acls.go
@@ -9,7 +9,7 @@ import (
 	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 
-	"github.com/fastly/go-fastly/v15/fastly/computeacls"
+	"github.com/fastly/go-fastly/v16/fastly/computeacls"
 	"github.com/fastly/terraform-provider-fastly/fastly/hashcode"
 )