+++ fastly/data_source_fastly_api_security_discovered_operations.go
@@ -10,8 +10,8 @@ import (
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/validation"
 
-	gofastly "github.com/fastly/go-fastly/v15/fastly"
-	"github.com/fastly/go-fastly/v15/fastly/apisecurity/operations"
+	gofastly "github.com/fastly/go-fastly/v16/fastly"
+	"github.com/fastly/go-fastly/v16/fastly/apisecurity/operations"
 	"github.com/fastly/terraform-provider-fastly/fastly/hashcode"
 )