+++ fastly/data_source_fastly_ngwaf_alert_slack_integration.go
@@ -11,7 +11,7 @@ import (
 
 	"github.com/fastly/terraform-provider-fastly/fastly/hashcode"
 
-	AlertSlackIntegrations "github.com/fastly/go-fastly/v15/fastly/ngwaf/v1/workspaces/alerts/slack"
+	AlertSlackIntegrations "github.com/fastly/go-fastly/v16/fastly/ngwaf/v1/workspaces/alerts/slack"
 )
 
 func dataSourceFastlyNGWAFAlertSlackIntegration() *schema.Resource {