+++ fastly/ngwaf_rule_flatten.go
@@ -5,8 +5,8 @@ import (
 
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 
-	"github.com/fastly/go-fastly/v15/fastly/ngwaf/v1/rules"
-	"github.com/fastly/go-fastly/v15/fastly/ngwaf/v1/scope"
+	"github.com/fastly/go-fastly/v16/fastly/ngwaf/v1/rules"
+	"github.com/fastly/go-fastly/v16/fastly/ngwaf/v1/scope"
 )
 
 func flattenNGWAFRuleResponse(d *schema.ResourceData, rule *rules.Rule) error {