+++ fastly/ngwaf_scope_helpers.go
@@ -7,8 +7,8 @@ import (
 
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 
-	gofastly "github.com/fastly/go-fastly/v15/fastly"
-	"github.com/fastly/go-fastly/v15/fastly/ngwaf/v1/scope"
+	gofastly "github.com/fastly/go-fastly/v16/fastly"
+	"github.com/fastly/go-fastly/v16/fastly/ngwaf/v1/scope"
 )
 
 type resolvedScope struct {