+++ fastly/resource_fastly_ngwaf_list.go
@@ -9,9 +9,9 @@ import (
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/validation"
 
-	gofastly "github.com/fastly/go-fastly/v15/fastly"
-	"github.com/fastly/go-fastly/v15/fastly/ngwaf/v1/lists"
-	"github.com/fastly/go-fastly/v15/fastly/ngwaf/v1/scope"
+	gofastly "github.com/fastly/go-fastly/v16/fastly"
+	"github.com/fastly/go-fastly/v16/fastly/ngwaf/v1/lists"
+	"github.com/fastly/go-fastly/v16/fastly/ngwaf/v1/scope"
 )
 
 func resourceFastlyNGWAFWorkspaceList() *schema.Resource {